Replace PickleSerializer with JSONSerializer
