+++ raiden/raiden_service.py
@@ -217,7 +217,7 @@ class RaidenService(Runnable):
             )
 
         # The database may be :memory:
-        storage = sqlite.SQLiteStorage(self.database_path, serialize.PickleSerializer())
+        storage = sqlite.SQLiteStorage(self.database_path, serialize.JSONSerializer())
         self.wal = wal.restore_from_latest_snapshot(
             node.state_transition,
             storage,
